overlap as float

+++ finetune.py
@@ -136,7 +136,7 @@ if __name__ == '__main__':
     parser.add_argument('--sample_generator', type = eval, nargs = '+', default = [10, 20, None])
     parser.add_argument('--standardize_channels', type = eval, default = True)
     parser.add_argument('--window_size', type = float, default = 0.2)
-    parser.add_argument('--overlap', type = int, default = 0.05)
+    parser.add_argument('--overlap', type = float, default = 0.05)
 
     # model arguments
     parser.add_argument('--remove_mpnn', type = eval, default = False)
